+++ plugins/psv-extractor/src/index.ts
@@ -1,2 +1,15 @@
-export * from "./psv.extractor";
-export * from "./psv.extractor.plugin";
+import { Flatfile } from '@flatfile/api'
+import { DelimiterExtractor } from '@flatfile/plugin-delimiter-extractor'
+
+export const PSVExtractor = (options?: {
+  dynamicTyping?: boolean
+  skipEmptyLines?: boolean | 'greedy'
+  transform?: (value: any) => Flatfile.CellValueUnion
+}) => {
+  return DelimiterExtractor('.psv', { delimiter: '|', ...options })
+}
+
+/*
+ * @deprecated use `PSVExtractor` instead
+ */
+export const psvExtractorPlugin = PSVExtractor