+++ plugins/record-hook/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @flatfile/plugin-record-hook
 
+## 1.0.1
+
+### Patch Changes
+
+- c859a10: Extract asyncBatch to a utility package
+- Updated dependencies [c859a10]
+  - @flatfile/util-common@0.0.2
+
+## 1.0.0
+
+### Major Changes
+
+- fe6e642: Create bulkRecordHook plugin
+
+## 0.2.0
+
+### Minor Changes
+
+- 2b15b32: Refactor extractors to share a common extractor and use custom parsers. Add .tsv file extractor and a generic delimiter extractor. Add header row detection to the Excel extractor.
+
+## 0.1.11
+
+### Patch Changes
+
+- 8a87451: Updates record-hook dependencies
+
 ## 0.1.10
 
 ### Patch Changes